requirements.txt
# Core
numpy==2.3.3
pandas==2.3.3
scikit-learn==1.7.2
xgboost==3.0.5
joblib==1.5.2

# API
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.9.2

# Visualization / analysis
matplotlib==3.10.0
seaborn==0.13.2
plotly==6.3.1

# EDA / notebooks (optional but common)
jupyterlab==4.3.1
ipykernel==6.29.5

# Optional explainability
shap==0.48.0

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
streamlit==1.39.0
